requirements.txt
Flask
gunicorn
Pillow
openai